refactor: NewCardScreen으로 분리 및 MainActivity 이름 복구
