+++ app/src/main/java/woowacourse/payments/presentation/newcard/MainActivity.kt
@@ -1,0 +1,19 @@
+package woowacourse.payments.presentation.newcard
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import woowacourse.payments.ui.theme.AndroidpaymentsTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            AndroidpaymentsTheme {
+                NewCardScreen()
+            }
+        }
+    }
+}
